adding changes related to disk mounting

+++ hdfs2gcs-terraform/nifi-ca.tf
@@ -41,24 +41,27 @@ resource "google_compute_instance" "nifi-ca" {
     }
 
     metadata_startup_script =   <<EOF
+        if [[ ! -f /opt/startup-script-finished.txt ]]
+        then
+            if [[ "${var.image}" == *"centos"* ]]; then
+                yum install     java-11-openjdk-devel unzip -y
+            else 
+                apt-get update && apt-get -yq install    openjdk-11-jdk unzip 
+            fi
+                    
+            NIFI_UID=10000
+            NIFI_GID=10000
 
-        if [[ "${var.image}" == *"centos"* ]]; then
-            yum install     java-11-openjdk-devel unzip -y
-        else 
-            apt-get update && apt-get -yq install    openjdk-11-jdk unzip 
+            groupadd -g $${NIFI_GID} nifi || groupmod -n nifi `getent group $${NIFI_GID} | cut -d: -f1` \
+                && useradd --shell /bin/bash -u $${NIFI_UID} -g $${NIFI_GID} -m nifi \
+                && mkdir -p ${var.nifi-path} \
+            
+            chown nifi:nifi ${var.nifi-path}/
+            su nifi -c 'curl -fSL https://archive.apache.org/dist/nifi/${var.nifi-version}/nifi-toolkit-${var.nifi-version}-bin.zip -o ${var.nifi-path}/nifi-toolkit-${var.nifi-version}-bin.zip'
+            su nifi -c 'unzip ${var.nifi-path}/nifi-toolkit-${var.nifi-version}-bin.zip -d ${var.nifi-path}'
+            su nifi -c 'rm ${var.nifi-path}/nifi-toolkit-${var.nifi-version}-bin.zip'
+            touch /opt/startup-script-finished.txt && echo "the startup script run once" > /opt/startup-script-finished.txt
         fi
-                
-        NIFI_UID=10000
-        NIFI_GID=10000
-
-        groupadd -g $${NIFI_GID} nifi || groupmod -n nifi `getent group $${NIFI_GID} | cut -d: -f1` \
-            && useradd --shell /bin/bash -u $${NIFI_UID} -g $${NIFI_GID} -m nifi \
-            && mkdir -p ${var.nifi-path} \
-         
-        chown nifi:nifi ${var.nifi-path}/
-        su nifi -c 'curl -fSL https://archive.apache.org/dist/nifi/${var.nifi-version}/nifi-toolkit-${var.nifi-version}-bin.zip -o ${var.nifi-path}/nifi-toolkit-${var.nifi-version}-bin.zip'
-        su nifi -c 'unzip ${var.nifi-path}/nifi-toolkit-${var.nifi-version}-bin.zip -d ${var.nifi-path}'
-        su nifi -c 'rm ${var.nifi-path}/nifi-toolkit-${var.nifi-version}-bin.zip'
         su nifi -c 'cd /home/nifi && ${var.nifi-path}/nifi-toolkit-${var.nifi-version}/bin/tls-toolkit.sh server -c ${var.nifi-ca-hostname} -t ${var.ca-token} &'
         sleep 10
         cd /root
